Attempt 5: fix CI builds, fix some warnings on MSVC

+++ src/core/lexer/Lexer.h
@@ -45,7 +45,7 @@ namespace core
 			const ContentIterator end;
 
 			SourceLocation currentLocation;
-			size_t lastLineLen;
+			uint64_t lastLineLen;
 
 			ErrorLevel error;
